% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{dickensdf}
\alias{dickensdf}
\title{Three sentences of dickens into a data frame}
\format{a data frame with 34 rows and 6 variables}
\usage{
dickensdf
}
\description{
Each row is a token. The variables are as follows:
}
\details{
\itemize{
  \item token : wordform
  \item pos : part of speach
  \item lemma : the citation form of the token
  \item sentence numeric : the id of the sentence
  \item np numeric : the id of the phrase
  \item doc_id : the id of the document
}

doc_id and token are two mandatory colums according to
the TIF (Text interchange format) specification (https://github.com/ropensci/tif)
}
\keyword{datasets}
