% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{dickensl}
\alias{dickensl}
\title{Three sentences of dickens into a list}
\format{a list of 3 slots containing character vector.}
\usage{
dickensl
}
\description{
Each slot is a character vector. The elements of the character vector contain the tokens.
}
\keyword{datasets}
