% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{dickensv}
\alias{dickensv}
\title{Three sentences of dickens into a character vector}
\format{a list of 3 slots containing character vector.}
\usage{
dickensv
}
\description{
Each element of the character vector contain an untokenized sentence.
}
\keyword{datasets}
