% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{kwic-package}
\alias{kwic-package}
\alias{_PACKAGE}
\title{Print Textual Corpora as Key Word in Context (aka concordance)}
\description{
Print Textual Corpora as Key Word in Context (aka concordance)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sylvainloiseau/kwic}
}

}
\author{
\strong{Maintainer}: Sylvain Loiseau \email{sylvain.loiseau@univ-paris13.fr}

}
