% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwic.R
\docType{methods}
\name{kwic}
\alias{kwic}
\alias{kwic,character-method}
\alias{kwic,list-method}
\alias{kwic,VCorpus-method}
\alias{kwic,data.frame-method}
\title{Create a kwic from vector, list, data.frame or other structure containing linguistic corpora}
\usage{
kwic(corpus, pattern, left = ifelse(unit == "char", 20, 5),
  right = ifelse(unit == "char", 20, 5), unit = "char", fixed = TRUE,
  ref = NULL, ...)

\S4method{kwic}{character}(corpus, pattern, left = 20, right = 20,
  unit = "char", fixed = TRUE, ref = names(corpus))

\S4method{kwic}{list}(corpus, pattern, left = 20, right = 20,
  unit = "char", fixed = TRUE, ref = names(corpus))

\S4method{kwic}{VCorpus}(corpus, pattern, left = 20, right = 20,
  unit = "char", fixed = TRUE, ref = names(corpus))

\S4method{kwic}{data.frame}(corpus, pattern, left = 5, right = 5,
  unit = "char", ref = NULL, token.column = "token",
  id.column = "doc_id", interlinearize.with = NULL)
}
\arguments{
\item{corpus}{the corpus (various data structure)}

\item{pattern}{length-1 character vector or either regexpr or fixed string to be search for}

\item{left}{length-1 integer vector : number of chars/tokens (see unit) on the right size}

\item{right}{length-1 integer vector : number of chars/tokens (see unit) on the left size}

\item{unit}{length-1 character vector : one of "char" or "token" : defines the left and right contexts as number of character or as number of words}

\item{fixed}{length-1 logical vector : is the pattern argument to be interpreted as a regexpr or as a fixed string}

\item{ref}{character vectors: the name for the different parts of the corpus}

\item{...}{unused arguments}

\item{token.column}{length-1 character vector : the name of the column containing the occurrences.
`token` is the default, according to Text Interchange Formats (see reference).}

\item{id.column}{length-1 character vector : the name column of the column for creating textual unit you don't wan't the kwic to cross.
`doc_id` is the default, as it is supposed
to exist in all data.frame according to Text Interchange Formats (see reference).}

\item{interlinearize.with}{character vector : the name of other column with which one can search.}
}
\value{
a KwicLine or KwicToken object (depending on the value of unit)
}
\description{
Create a kwic from vector, list, data.frame or other structure containing linguistic corpora
}
\examples{
# Concordance with a vector of untokenized strings
data(dickensv)
kwic(dickensv, "the")
# Concordance with a list of tokens
data(dickensl)
kwic(dickensl, "the")

# Concordance with a tm object
library(tm)
data(acq)
kwic(acq, "stock")
# Concordance with a data frame. Defaults are used for the arguments
# 'token.column' 'id.column' (ie column names 'token' and 'doc_id')
data(dickensdf)
kwic(dickensdf, "the")
}
\references{
Text Interchange Formats : https://github.com/ropensci/tif
}
