% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicLine.R
\docType{class}
\name{KwicLine-class}
\alias{KwicLine-class}
\title{A class representing an extraction of kwic as raw character sequences from a corpora, ready to be printed.}
\description{
A class representing an extraction of kwic as raw character sequences from a corpora, ready to be printed.
}
\section{Slots}{

\describe{
\item{\code{ref}}{character.}

\item{\code{left.context.slot}}{character.}

\item{\code{right.context.slot}}{character.}

\item{\code{keyword.slot}}{character.}

\item{\code{keyword}}{character.}

\item{\code{leftspan}}{numeric.}

\item{\code{rightspan}}{numeric.}

\item{\code{nbr.match}}{numeric.}

\item{\code{nbr.match.by.line}}{numeric.}

\item{\code{keyword.match.length}}{numeric.}
}}

\examples{
data(dickensl, unit="line")
kwic(dickensl, "the")
}
