% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicToken.R
\docType{class}
\name{KwicToken-class}
\alias{KwicToken-class}
\title{#' A class representing an extraction of kwic as a sequences of tokens from a corpora, ready to be printed.}
\description{
#' A class representing an extraction of kwic as a sequences of tokens from a corpora, ready to be printed.
}
\section{Slots}{

\describe{
\item{\code{ref}}{character.}

\item{\code{concordances.matrix}}{matrix.}

\item{\code{keyword}}{character.}

\item{\code{leftspan}}{numeric.}

\item{\code{rightspan}}{numeric.}

\item{\code{nbr.match}}{numeric.}

\item{\code{nbr.match.by.line}}{numeric.}

\item{\code{keyword.match.length}}{numeric.}
}}

\examples{
data(dickensl)
kwic(dickensl, "the", unit="tokens")
}
