% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicLine.R
\docType{methods}
\name{print,KwicLine-method}
\alias{print,KwicLine-method}
\title{Print a KwicLine object}
\usage{
\S4method{print}{KwicLine}(x, from = 1, to = -1, sort.by = "none",
  decreasing = FALSE, file = "", append = FALSE, ...)
}
\arguments{
\item{x}{KwicLine object: the set of kwic lines}

\item{from}{length-1 numeric vector: index of the first line to be printed}

\item{to}{length-1 numeric vector: index of the last line to be printed}

\item{sort.by}{length-1 character vector: sort lines by 'right' context or 'left' context ('none': corpus order, 'random': randomized order)}

\item{decreasing}{lenth-1 logical vector: is the sort to be done in increasing or decreasing order?}

\item{file}{length-1 character vector: url of the file to be used for writing the kwic lines}

\item{append}{length-1 logical vector: should the content of the file be erased or preserved}

\item{...}{unused arguments}
}
\description{
Print a KwicLine object
}
\examples{
data(dickensl, unit="line")
res <- kwic(dickensl, "the")
print(res)
}
