% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicWord.R
\docType{methods}
\name{print,KwicWord-method}
\alias{print,KwicWord-method}
\title{Print a KwicWord object}
\usage{
\S4method{print}{KwicWord}(x, from = 1, to = -1, sort.by = "none",
  decreasing = FALSE, file = "", append = FALSE, ...)
}
\arguments{
\item{x}{KwicWord object: the set of kwic lines}

\item{from}{length-1 numeric vector: index of the first line to be printed}

\item{to}{length-1 numeric vector: index of the last line to be printed}

\item{sort.by}{length-1 character or numeric vector: sorting lines. 'none': corpus order; 'right' and 'left': first word after or before the keyword; 'random': randomized order; 0: keyword, n < 0 : the n^{th} word before the keyword; n > 0 : the n^{th} word after the keyword}

\item{decreasing}{lenth-1 logical vector: is the sort to be done in increasing or decreasing order?}

\item{file}{length-1 character vector: url of the file to be used for writing the kwic lines}

\item{append}{length-1 logical vector: should the content of the file be erased or preserved}

\item{...}{unused arguments}
}
\description{
Print a KwicWord object
}
\examples{
data(dickensl, unit="token")
res <- kwic(dickensl, "the")
print(res)
}
