% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicLine.R
\docType{methods}
\name{show,KwicLine-method}
\alias{show,KwicLine-method}
\title{Show a KwicLine object}
\usage{
\S4method{show}{KwicLine}(object)
}
\arguments{
\item{object}{KwicLine.}
}
\description{
Show a KwicLine object
}
\examples{
data(dickensl, unit="line")
res <- kwic(dickensl, "the")
show(res)
}
