% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicToken.R
\docType{methods}
\name{show,KwicToken-method}
\alias{show,KwicToken-method}
\title{Show a KwicToken object}
\usage{
\S4method{show}{KwicToken}(object)
}
\arguments{
\item{object}{KwicToken.}
}
\description{
Show a KwicToken object
}
\examples{
data(dickensl, unit="token")
res <- kwic(dickensl, "the")
show(res)
}
