% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicWord.R
\docType{methods}
\name{show,KwicWord-method}
\alias{show,KwicWord-method}
\title{Show a KwicWord object}
\usage{
\S4method{show}{KwicWord}(object)
}
\arguments{
\item{object}{KwicWord.}
}
\description{
Show a KwicWord object
}
\examples{
data(dickensl, unit="token")
res <- kwic(dickensl, "the")
show(res)
}
