% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicLine.R
\docType{methods}
\name{summary,KwicLine-method}
\alias{summary,KwicLine-method}
\title{Print a summary of a KwicLine object.}
\usage{
\S4method{summary}{KwicLine}(object)
}
\arguments{
\item{object}{KwicLine.}
}
\description{
Print a summary of a KwicLine object.
}
\examples{
data(dickensl, unit="line")
res <- kwic(dickensl, "the")
summary(res)
}
