% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicToken.R
\docType{methods}
\name{summary,KwicToken-method}
\alias{summary,KwicToken-method}
\title{Print a summary of a KwicToken object}
\usage{
\S4method{summary}{KwicToken}(object)
}
\arguments{
\item{object}{KwicToken.}
}
\description{
Print a summary of a KwicToken object
}
\examples{
data(dickensl, unit="token")
res <- kwic(dickensl, "the")
summary(res)
}
