% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KwicWord.R
\docType{methods}
\name{summary,KwicWord-method}
\alias{summary,KwicWord-method}
\title{Print a summary of a KwicWord object}
\usage{
\S4method{summary}{KwicWord}(object)
}
\arguments{
\item{object}{KwicWord.}
}
\description{
Print a summary of a KwicWord object
}
\examples{
data(dickensl, unit="token")
res <- kwic(dickensl, "the")
summary(res)
}
